Use NS_ENUM for better Swift interoperability
Objective-C enums that are defined with the NS_ENUM macro are automatically imported
into Swift as Swift enumerations. This allows for easier usage of the enums in Swift.
In addition to wrapping all enum declarations into NS_ENUM, this commit also makes
sure that the correct factory method is used when converting between enum constants
and NSNumbers.