Use external linkage & const pointers for shared string constants and make unknown region code public
The shared string constants in NBPhoneNumberDefines.h were defined as static variables.
As a result, they were not actually shared but instead a new instance was created for
every file that imported the header. This commit changes the constants to use external
linkage which makes them truly shared. On top, the pointers were made immutable. Finally,
the special @"ZZ" code for unknown regions was added to the public shared string constants.
Since this code can be returned by a number of public library functions, clients should
have a way to check for it other than using the internal string literal @"ZZ".