Remove deprecated call to RunUnitTests and invalid/superfluous search paths
This commit removes the run script build phase from the test target which called
the RunUnitTests command. The latter is obsolete now since unit test execution
is configured in schemes. Removing the build phase removes a deprecation warning
that appeared when running the tests in Xcode.

In addition, this commit drops invalid and superfluous search paths from the
project file. An invalid framework search path caused a warning when running
the tests in Xcode.